refactor(wallet-dashboard): move state to page
