+++ apps/wallet-dashboard/components/Dialogs/Assets/constants/AssetsDialogView.ts
@@ -1,0 +1,7 @@
+// Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+export enum AssetsDialogView {
+    Details = 'Details',
+    Send = 'Send',
+}
