--- apps/wallet-dashboard/components/Dialogs/Assets/hooks/useAssetsDialog.ts
@@ -1,18 +1,0 @@
-// Copyright (c) 2024 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-import React from 'react';
-
-export enum AssetsDialogView {
-    Details = 'Details',
-    Send = 'Send',
-}
-
-export function useAssetsDialog() {
-    const [view, setView] = React.useState<AssetsDialogView | undefined>(AssetsDialogView.Details);
-
-    return {
-        view,
-        setView,
-    };
-}
